feat: added decimal formatting for quote and quantity fields for appropriate inputs and added pending payments check for requesting amount, added new vendor addition sheet for inprogress prs and sbs, chores: fixed ui bugs
