+++ frontend/src/components/helpers/SelectUnit.tsx
@@ -1,0 +1,47 @@
+import { Select, SelectContent, SelectItem, SelectTrigger, SelectValue } from "../ui/select";
+
+interface SelectUnitProps {
+    value: string;
+    onChange: (value: string) => void;
+    className?: string;
+    disabled?: boolean;
+}
+
+export const SelectUnit : React.FC<SelectUnitProps> = ({ value, onChange, className, disabled = false }) => {
+
+  const options : {value: string, label: string}[] = [
+    // { value: "PCS", label: "PCS" },
+    // {value : "PKT", label : "PKT"},
+    { value: "BOX", label: "BOX" },
+    { value: "ROLL", label: "ROLL" },
+    {value : "LTH", label : "LTH"},
+    {value : "MTR", label : "MTR"},
+    { value: "NOS", label: "NOS" },
+    { value: "KGS", label: "KGS" },
+    { value: "PAIRS", label: "PAIRS" },
+    { value: "PACKS", label: "PACKS" },
+    { value: "DRUM", label: "DRUM" },
+    { value: "SQMTR", label: "SQMTR" },
+    { value: "LTR", label: "LTR" },
+    { value: "BUNDLE", label: "BUNDLE" },
+    { value: "FEET", label: "FEET" },
+  ]
+    return (
+        <Select
+            value={value}
+            onValueChange={(e) => onChange(e)}
+            disabled={disabled}
+        >
+          <SelectTrigger>
+            <SelectValue className={className} placeholder="Select Unit" />
+          </SelectTrigger>
+          <SelectContent>
+            {options.map((option) => (
+                <SelectItem key={option.value} value={option.value}>
+                    {option.label}
+                </SelectItem>
+            ))}
+          </SelectContent>
+        </Select>
+    );
+};
